Changed the home screen
So now instead of the two send buttons it is a single add button in the bottom right hand corner. Currently a email icon havent changed it but after pressing it will take you to a screen which would resemble the activity management mode. The TextView string is updated to the input of the two text fields concatinated together when the create button is pressed.  This is a pure visual representation of string maniplation which is how we will interact with the database. Currently still no connections set up properly still in dev phase.